feat: add choosy

+++ .config/install.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
 # https://www.iterm2.com/documentation-shell-integration.html
-curl -L https://iterm2.com/shell_integration/zsh -o ~/.iterm2_shell_integration.zsh
+if [ ! -f ~/.iterm2_shell_integration.zsh ]; then
+    echo "Installing iTerm2 shell integration"
+    curl -L https://iterm2.com/shell_integration/zsh -o ~/.iterm2_shell_integration.zsh
+fi
 
 # https://brew.sh/
 (brew --version > /dev/null) || (echo "Installing brew" && /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)")
@@ -19,4 +22,20 @@ fi
 if [ ! -d "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k" ]; then
     echo "Installing powerlevel10k theme"
     git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+fi
+
+# https://www.choosyosx.com/
+if [ ! -L ~/Library/Application\ Support/Choosy/behaviours.plist ]; then
+    if [ -d ~/Library/Application\ Support/Choosy ]; then
+        echo "Installing Choosy profile"
+        mv ~/Library/Application\ Support/Choosy/behaviours.plist ~/Library/Application\ Support/Choosy/behaviours.bak.plist
+        config_dir="$(dirname $0)"
+        config_dir="$(realpath $config_dir)"
+        ln -s "${config_dir}/choosy/behaviours.plist" ~/Library/Application\ Support/Choosy/behaviours.plist
+        echo "Restarting Choosy"
+        killall Choosy
+        open /Applications/Choosy.app
+    else
+        echo "Please install Choosy from https://www.choosyosx.com/, launch it and retry this script"
+    fi
 fi
